Support atom links in RSS.
The end goal is that Atom link elements in RSS feeds work. However in order to achieve that a major refactoring to how namespace work was required.

Previously Atom and RSS namespace were mapped to `None` so that feeds missing a top-level namespace implicitly matched. This was a clever hack but made it impossible to identify Atom elements in RSS feeds and vice-versa.

In order to fix this we now always have a namespace, so instead of `Option<NS>` we pass around `NS`. In order to deal with the default namespace we simply set it based on the feed type. So if a feed is RSS2 we set the default namespace to RSS. This keeps all parser code from having to know if the namespace is explicit or defaulted.

Fixes https://github.com/feed-rs/feed-rs/issues/137