Copy entry.published to entry.updated for RSS2
Previously, the 'updated' field of an entry was populated from the
'updated' field on a feed.

However this caused all entries within a feed to be reported as updated
each time the parent feed changed, which is incorrect.

The RSS2 parser now copies across the entry.published field to
entry.updated instead, maintaining the use case of the 'updated' field
by avoiding spurious reports of updated entries.

Fixes #235.