Add support for the `src` attribute for `atom:content`.
This is described in https://datatracker.ietf.org/doc/html/rfc4287#section-4.1.3.2

Unfortunately the existing model types are a bit at odds with the spec. To avoid breaking backwards compatibility I had to be a bit creative with the types.

1. If the `content_type` is not provided we guess that it is `text/html` as the `Content` type requires a type. This is likely not an issue because the `body` is `None`.
2. There is no appropriate error value for having both content and src.

Both can be fixed in a future version when backwards compatibility is broken.

For 2 it may also make sense to just return `src` and `body`. The code could treat them mostly independently. For now it is an error, which can always be changed if this is seen in the wild.